Introduce MountedFast function to determine mount-points faster.
MountedFast is a method of detecting mount points ie without reading
/proc/mounts. It works for all kinds of mounts (incl. bind mounts)
but requires a recent (v5.6+) Linux kernel. A mount-point check is
guaranteed to be a mount-point or not only when sure is true. When
sure is false, the caller needs to check for other methods
(eg: parse /proc/mounts) to successfully determine if it is a
mount-point.